Adding parallel property-based tests
Sadly the parallel tests require to lower the number of resources
to one given the non-determinism of the default behaviour appears
to conflict with the interleaving attempts of PropEr.

Mocking the dispatch_id function of dispcount_watcher appeared
to have little effect. The test is a best-effort for now;
better than nothing, although not perfect.