Properly re-add the watchers_sup to the dispcount_sup on restart
If the dispcount_serv dies (or is restarted because of too many worker
failures), the addition of the watchers_sup to the supervisor tree may
fail because an old child spec (with invalid references) is still
installed. This change ensures that if an old supervisor is still
installed it is terminated and deleted before adding a new one.